requirements.txt
torch
transformers
pillow
requests
torchvision
matplotlib
kagglehub
streamlit